requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
beautifulsoup4
faiss-cpu
python-dotenv
tiktoken